JSCOE-155 initial voting components
