+++ eventsource-eventbridge-lambda/audit/README.md
@@ -1,0 +1,13 @@
+# Audit Service
+
+This project provisions Lambda function(s), and EventBridge resources to provide _Election Audit_ domain functionality.
+
+This project uses the Serverless Framework.
+
+## Install
+
+## Run
+
+## Remove / Clean Up
+
+## Related Information
