+++ eventsource-eventbridge-lambda/audit/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "jscoe155-notifications",
+  "name": "jscoe155-audit",
   "version": "1.0.0",
-  "description": "JSCOE-155 - Event sourcing pattern experiment with Amazon EventBridge and Lambda. Notifications component.",
+  "description": "JSCOE-155 - Event sourcing pattern experiment with Amazon EventBridge and Lambda. Audit component.",
   "main": "index.js",
   "repository": "git@github.com:ModusCreateOrg/js-coe-experiments.git",
   "author": "Matt Warman",
