--- eventsource-eventbridge-lambda/books/README.md
@@ -1,13 +1,0 @@
-# Book Service
-
-This project provisions Lambda function(s), and EventBridge resources to provide _Book_ domain functionality.
-
-This project uses the Serverless Framework.
-
-## Install
-
-## Run
-
-## Remove / Clean Up
-
-## Related Information
