+++ eventsource-eventbridge-lambda/elections/README.md
@@ -1,0 +1,13 @@
+# Election Service
+
+This project provisions Lambda function(s), and EventBridge resources to provide _Election_ domain functionality.
+
+This project uses the Serverless Framework.
+
+## Install
+
+## Run
+
+## Remove / Clean Up
+
+## Related Information
