+++ eventsource-eventbridge-lambda/elections/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "jscoe155-books",
+  "name": "jscoe155-elections",
   "version": "1.0.0",
-  "description": "JSCOE-155 - Event sourcing pattern experiment with Amazon EventBridge and Lambda. Books component.",
+  "description": "JSCOE-155 - Event sourcing pattern experiment with Amazon EventBridge and Lambda. Election component.",
   "main": "index.js",
   "repository": "git@github.com:ModusCreateOrg/js-coe-experiments.git",
   "author": "Matt Warman",
@@ -16,6 +16,5 @@
     "serverless": "3.1.1"
   },
   "dependencies": {
-    "axios": "0.25.0"
   }
 }
