+++ eventsource-eventbridge-lambda/electionsummary/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "jscoe155-electionsummary",
+  "version": "1.0.0",
+  "description": "JSCOE-155 - Event sourcing pattern experiment with Amazon EventBridge and Lambda. Election Summary component.",
+  "main": "index.js",
+  "repository": "git@github.com:ModusCreateOrg/js-coe-experiments.git",
+  "author": "Matt Warman",
+  "license": "UNLICENSED",
+  "private": true,
+  "scripts": {
+    "deploy": "sls deploy",
+    "logs": "sls logs --tail",
+    "remove": "sls remove"
+  },
+  "devDependencies": {
+    "serverless": "3.1.1"
+  },
+  "dependencies": {
+  }
+}
