--- eventsource-eventbridge-lambda/notifications/README.md
@@ -1,13 +1,0 @@
-# Book Service
-
-This project provisions Lambda function(s), and EventBridge resources to provide _Notification_ domain functionality.
-
-This project uses the Serverless Framework.
-
-## Install
-
-## Run
-
-## Remove / Clean Up
-
-## Related Information
