+++ eventsource-eventbridge-lambda/voters/serverless.yml
@@ -1,0 +1,48 @@
+service: jscoe155-voters
+
+custom:
+  eventBusStack: 'jscoe155-events'
+  eventSource: 'moduscreate.jscoe155'
+  stage: ${opt:stage, self:provider.stage}
+  tags:
+    App: 'jscoe.moduscreate.com'
+    Env: ${self:custom.stage}
+    OU: 'moduscreate'
+    Owner: 'Matthew Warman'
+    Issue: 'JSCOE-155'
+
+provider:
+  name: aws
+  stage: dev
+  region: us-east-1
+  runtime: nodejs14.x
+  memorySize: 128
+  timeout: 6
+  environment:
+    EVENTBRIDGE_SOURCE: ${self:custom.eventSource}
+    API_BASEURL: ${cf:${self:custom.eventBusStack}-${self:custom.stage}.ServiceEndpoint}
+  logRetentionInDays: 7
+  tags: ${self:custom.tags}
+
+functions:
+  ballotSubmitted:
+    handler: src/handlers/ballotSubmitted.handle
+    events:
+      - eventBridge:
+          eventBus: ${cf:${self:custom.eventBusStack}-${self:custom.stage}.EventBusArn}
+          pattern:
+            source:
+              - ${self:custom.eventSource}
+            detail-type:
+              - 'Ballot Submitted'
+          deadLetterQueueArn: !GetAtt BallotSubmittedRuleDLQ.Arn
+          retryPolicy:
+            maximumEventAge: 86400 # default 1 day
+            maximumRetryAttempts: 185 # default 185 attempts
+
+resources:
+  Resources:
+    BallotSubmittedRuleDLQ:
+      Type: AWS::SQS::Queue
+      Properties:
+        QueueName: ${self:service}-${self:custom.stage}-ballotSubmittedRule-DLQ
